feat: refactor in ECMAScript notation modules
